Split components into 2
